Do not remove decrypted secrets
Information becomes visible to root anyways, so storing it with root
only access is good enough.

This will make it easier on bootstrap, when yubikey/gpg related services
are not ready yet, by keeping the decrypted secrets from the live iso
session.

It will still be easy to keep secrets in one place, since the encrypted
version will overwrite the previous decrypted version on each run.